More fixes

+++ w3af/plugins/attack/payloads/payloads/tcp.py
@@ -39,7 +39,7 @@ class tcp(Payload):
             while d > 0:
                 m, n = divmod(n, d)
                 q.append(str(m))
-                d = d / 256
+                d = int(d / 256)
             q.reverse()
             return '.'.join(q)
